Move CEPGetterEndpoint to its own file
